+++ tests/examples/KWK_process.py
@@ -116,7 +116,7 @@ for current_station in ['HOEKVHLD','DENOVBTN']:#stat_list: #stat_list[stat_list.
         data_pd_HWLW_all = clean_data(data_pd_HWLW_all,current_station)
         if compute_slotgem or compute_havengetallen or compute_overschrijding: #TODO: make calc_HWLW12345to12() faster
             data_pd_HWLW_all_12 = hatyan.calc_HWLW12345to12(data_pd_HWLW_all) #convert 12345 to 12 by taking minimum of 345 as 2 (laagste laagwater)
-            #crop timeseries to 10y
+            #crop timeseries to 10y]
             data_pd_HWLW_10y_12 = hatyan.crop_timeseries(data_pd_HWLW_all_12, times_ext=[tstart_dt,tstop_dt],onlyfull=False)
             
             #check if amount of HWs is enough