+++ tests/test_timeseries.py
@@ -11,6 +11,7 @@ import pandas as pd
 import datetime as dt
 import numpy as np
 import hatyan
+from hatyan.metadata import metadata_from_obj, metadata_compare
 
 dir_tests = os.path.dirname(__file__) #F9 doesnt work, only F5 (F5 also only method to reload external definition scripts)
 dir_testdata = os.path.join(dir_tests,'data_unitsystemtests')
@@ -160,3 +161,81 @@ def test_readts_dia_equidistant_multifile_glob_hasfreq():
     assert isinstance(ts_pd.index.freq,pd.offsets.Minute)
     assert ts_pd.index.freq is not None
     assert ts_pd.index.freq.nanos/1e9 == 3600
+
+
+@pytest.mark.unittest
+def test_crop_timeseries():
+    
+    current_station = 'VLISSGN'
+    times_ext=[dt.datetime(2019,1,1),dt.datetime(2019,6,1)]
+    
+    file_pred = os.path.join(dir_testdata,f'{current_station}_pre.txt')
+    ts_prediction = hatyan.readts_dia(filename=file_pred, station=current_station)
+    ts_prediction_cropped = hatyan.crop_timeseries(ts_prediction, times_ext=times_ext)
+    
+    assert len(ts_prediction_cropped) == 21745
+    assert ts_prediction_cropped.index[0] == pd.Timestamp(times_ext[0])
+    assert ts_prediction_cropped.index[-1] == pd.Timestamp(times_ext[-1])
+    
+    pred_meta = metadata_from_obj(ts_prediction)
+    pred_cropped_meta = metadata_from_obj(ts_prediction_cropped)
+    metadata_compare([pred_meta,pred_cropped_meta])
+    
+    assert pred_cropped_meta['tstart'] == pd.Timestamp(times_ext[0])
+    assert pred_cropped_meta['tstop'] == pd.Timestamp(times_ext[-1])
+
+
+
+@pytest.mark.unittest
+def test_resample_timeseries():
+    
+    current_station = 'VLISSGN'
+    timestep_min = 120
+    
+    file_pred = os.path.join(dir_testdata,f'{current_station}_pre.txt')
+    ts_prediction = hatyan.readts_dia(filename=file_pred, station=current_station)
+    ts_prediction_res = hatyan.resample_timeseries(ts_prediction, timestep_min=timestep_min)
+    
+    assert len(ts_prediction_res) == 4380
+    assert ts_prediction_res.index[0] == pd.Timestamp(ts_prediction.index[0])
+    assert ts_prediction_res.index[-1] == pd.Timestamp("2019-12-31 22:00")
+    
+    pred_meta = metadata_from_obj(ts_prediction)
+    pred_meta.pop('timestep_min')
+    pred_res_meta = metadata_from_obj(ts_prediction_res)
+    pred_res_meta.pop('timestep_min')
+    metadata_compare([pred_meta,pred_res_meta])
+    
+    pred_res_meta = metadata_from_obj(ts_prediction_res)
+    assert pred_res_meta['timestep_min'] == timestep_min
+
+
+@pytest.mark.unittest
+def test_write_tsdia_rounding():
+    """
+    rounding error occurred in older versions of hatyan2
+    Therefore we test whether writing and reading a timeseries results in the same data (accuracy of 1cm)
+    """
+    
+    current_station = 'HOEKVHLD'
+    file_data_comp0 = os.path.join(dir_testdata,f'{current_station}_ana.txt')
+    
+    times_ext_pred = [dt.datetime(2019,1,1),dt.datetime(2020,1,1)]
+    times_step_pred = 10
+    
+    comp_merged = hatyan.read_components(filename=file_data_comp0)
+    
+    #prediction and validation
+    ts_prediction = hatyan.prediction(comp=comp_merged, nodalfactors=True, xfac=True, fu_alltimes=False, 
+                                      times_ext=times_ext_pred, timestep_min=times_step_pred)
+    
+    #write to file
+    fname_pred = 'prediction_%im_%s.dia'%(times_step_pred,current_station)
+    hatyan.write_tsdia(ts=ts_prediction, filename=fname_pred)
+    
+    #read from file
+    ts_prediction_fromfile = hatyan.readts_dia(filename=fname_pred, station=current_station)
+    
+    # assert max differences
+    ts_diff = ts_prediction_fromfile -ts_prediction
+    assert (ts_diff['values'].abs()<=0.005).all()